refactor: migrate auth logics into a auth module
