--- bandlog-adapters/bandlog-user-adapter/build.gradle.kts
@@ -1,20 +1,0 @@
-plugins {
-    id("org.springframework.boot")
-    id("io.spring.dependency-management")
-}
-
-dependencies {
-    implementation(project(":bandlog-shared"))
-    implementation(project(":bandlog-user"))
-    implementation(project(":bandlog-ports:bandlog-user-port"))
-    implementation("org.springframework.boot:spring-boot-starter-web")
-    testImplementation("org.springframework.boot:spring-boot-starter-test")
-}
-
-tasks.named<org.springframework.boot.gradle.tasks.bundling.BootJar>("bootJar") {
-    enabled = false
-}
-
-tasks.named<Jar>("jar") {
-    enabled = true
-}
