+++ bandlog-auth/src/main/java/net/effize/bandlog/auth/exception/IllegalAuthenticationException.java
