+++ bandlog-auth/src/main/java/net/effize/bandlog/auth/exception/UserNotSignedUpException.java
