+++ bandlog-auth/src/main/java/net/effize/bandlog/auth/service/AuthService.java
@@ -3,11 +3,9 @@ package net.effize.bandlog.auth.service;
 import net.effize.bandlog.auth.exception.IllegalAuthenticationException;
 import net.effize.bandlog.auth.exception.UserNotSignedUpException;
 import net.effize.bandlog.auth.model.SupabaseAuthenticationPrincipal;
+import net.effize.bandlog.port.user.BandlogAuthUserPort;
+import net.effize.bandlog.port.user.dto.BandlogUserResponse;
 import net.effize.bandlog.shared.auth.AuthUser;
-import net.effize.bandlog.user.dto.response.UserResponse;
-import net.effize.bandlog.user.model.Email;
-import net.effize.bandlog.user.model.SupabaseUserId;
-import net.effize.bandlog.user.service.UserService;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -15,10 +13,10 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 public class AuthService {
 
-    private final UserService userService;
+    private final BandlogAuthUserPort authUserPort;
 
-    public AuthService(UserService userService) {
-        this.userService = userService;
+    public AuthService(BandlogAuthUserPort authUserPort) {
+        this.authUserPort = authUserPort;
     }
 
     @Transactional
@@ -29,7 +27,7 @@ public class AuthService {
             throw new IllegalAuthenticationException();
         }
 
-        userService.signUp(new SupabaseUserId(supabaseUserId), new Email(email));
+        authUserPort.signUp(supabaseUserId, email);
     }
 
     @Transactional(readOnly = true)
@@ -42,14 +40,14 @@ public class AuthService {
             throw new IllegalAuthenticationException();
         }
 
-        UserResponse user = userService.findBySupabaseUserId(new SupabaseUserId(authPrincipal.supabaseUserId()))
+        BandlogUserResponse userResponse = authUserPort.findBySupabaseUserId(authPrincipal.supabaseUserId())
                 .orElseThrow(UserNotSignedUpException::new);
 
         return new AuthUser(
-                user.id(),
-                user.supabaseUserId(),
-                user.email(),
-                user.nickname()
+                userResponse.id(),
+                userResponse.supabaseUserId(),
+                userResponse.email(),
+                userResponse.nickname()
         );
     }
 }
