+++ bandlog-bootstrap/build.gradle.kts
@@ -6,11 +6,9 @@ plugins {
 dependencies {
     implementation(project(":bandlog-shared"))
     implementation(project(":bandlog-user"))
+    implementation(project(":bandlog-auth"))
     implementation(project(":bandlog-team"))
-    implementation(project(":bandlog-adapters:bandlog-user-adapter"))
     implementation("org.springframework.boot:spring-boot-starter-web")
-    implementation("org.springframework.boot:spring-boot-starter-security")
-    implementation("org.springframework.boot:spring-boot-starter-oauth2-resource-server")
     implementation("org.springframework.boot:spring-boot-starter-data-jpa")
     runtimeOnly("org.postgresql:postgresql")
     testImplementation("org.springframework.boot:spring-boot-starter-test")
