+++ bandlog-ports/bandlog-user-port/src/main/java/net/effize/bandlog/port/user/BandlogAuthUserPort.java
@@ -1,0 +1,11 @@
+package net.effize.bandlog.port.user;
+
+import net.effize.bandlog.port.user.dto.BandlogUserResponse;
+
+import java.util.Optional;
+
+public interface BandlogAuthUserPort {
+    void signUp(String supabaseUserId, String email);
+
+    Optional<BandlogUserResponse> findBySupabaseUserId(String supabaseUserId);
+}
