+++ settings.gradle.kts
@@ -1,10 +1,9 @@
 rootProject.name = "bandlog"
 
-include("bandlog-user")
 include("bandlog-shared")
+include("bandlog-user")
+include("bandlog-auth")
 include("bandlog-team")
-include("bandlog-bootstrap")
-include("bandlog-adapters:bandlog-user-adapter")
 include("bandlog-ports")
 include("bandlog-ports:bandlog-user-port")
-include("bandlog-auth")
+include("bandlog-bootstrap")
